refactor(types): group provider/speech types under common/types/provider
