+++ packages/desktop/src/common/config/configKeys.ts
@@ -1,5 +1,5 @@
 import type { AcpInitializeResult, AcpSessionConfigOption, AcpSessionModes } from '@/common/types/acpTypes';
-import type { SpeechToTextConfig } from '@/common/types/speech';
+import type { SpeechToTextConfig } from '@/common/types/provider/speech';
 import type { ICssTheme, IMcpServer, TProviderWithModel } from '@/common/config/storage';
 
 export type ConfigKeyMap = {
