+++ packages/desktop/src/common/types/provider/speech.ts
